Improvements!
- It looks like I failed to upload a couple of changes to GitHub before
sending it over, which means the "full size carousel" template breaks on
mobile. I've made sure that goes in.
- I fixed a bug where the images on the 3 columns template weren't
responsive at all on Firefox or some versions of IE. You'll need to
append the class "img-column" to those images to get them to play nice.
- I fixed an issue where the full-size sections were slightly bigger
than the content area. It should play nice on most browsers now, but
something is still causing a horizontal scroll bar to appear on Firefox.